musicapp/requirements.txt
Django==4.2
djangorestframework==3.14.0
mysql-connector-python==8.0.32
python-dotenv==0.21.1
requests==2.28.2
